requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.64
langchain_openai==0.3.21
langgraph==0.4.8
pymongo==4.13.0
python-dotenv==1.1.0
Requests==2.32.4
